pyproject.toml
[project]
name = "newsprisma"
version = "0.1.0"
description = "Multilingual news RAG agent surfacing how the same story looks through EN/ES/RU media"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # ingestion
    "feedparser>=6.0.11",
    "trafilatura>=1.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    # dev / testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/newsprisma"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
